requirements.txt
apache-beam==2.28.0
apipkg==1.5
atomicwrites==1.4.0
attrs==20.3.0
avro-python3==1.9.2.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
crcmod==1.7
deprecation==2.1.0
dill==0.3.1.1
docker==4.4.4
docopt==0.6.2
execnet==1.8.0
fastavro==1.3.4
fasteners==0.16
freezegun==1.1.0
future==0.18.2
google-api-core==1.26.2
google-apitools==0.5.31
google-auth==1.28.0
google-cloud-bigquery==1.28.0
google-cloud-bigtable==1.7.0
google-cloud-build==2.0.0
google-cloud-core==1.6.0
google-cloud-datastore==1.15.3
google-cloud-dlp==1.0.0
google-cloud-language==1.3.0
google-cloud-pubsub==1.7.0
google-cloud-spanner==1.19.1
google-cloud-videointelligence==1.16.1
google-cloud-vision==1.0.0
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
greenlet==1.0.0
grpc-google-iam-v1==0.12.3
grpcio==1.36.1
grpcio-gcp==0.2.2
hdfs==2.6.0
httplib2==0.17.4
idna==2.10
importlib-metadata==3.7.3
libcst==0.3.17
mock==2.0.0
more-itertools==8.7.0
mypy-extensions==0.4.3
nose==1.3.7
nose-xunitmp==0.4.1
numpy==1.19.5
oauth2client==4.1.3
packaging==20.9
pandas==1.1.5
parameterized==0.7.5
pbr==5.5.1
pluggy==0.13.1
proto-plus==1.18.1
protobuf==3.15.6
psycopg2-binary==2.8.6
py==1.10.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.2
PyHamcrest==1.10.1
pymongo==3.11.3
pyparsing==2.4.7
pytest==4.6.11
pytest-cov==2.11.1
pytest-forked==1.3.0
pytest-timeout==1.4.2
pytest-xdist==1.34.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-mock==1.8.0
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.2
tenacity==5.1.5
testcontainers==3.3.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.4
wcwidth==0.2.5
websocket-client==0.58.0
wrapt==1.12.1
zipp==3.4.1
